fix: Update setup file structure
