+++ setup/uninstall.py
